+++ pinot-broker/src/main/java/org/apache/pinot/broker/grpc/GrpcRequesterIdentity.java
@@ -21,7 +21,7 @@ package org.apache.pinot.broker.grpc;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
 import java.util.Map;
-import org.apache.pinot.broker.api.RequesterIdentity;
+import org.apache.pinot.spi.auth.broker.RequesterIdentity;
 import org.apache.pinot.common.proto.Broker;
 import org.apache.pinot.common.proto.Server;
 import org.apache.pinot.spi.utils.CommonConstants;
