+++ pinot-broker/src/test/java/org/apache/pinot/broker/api/AccessControlBackwardCompatibleTest.java
@@ -21,6 +21,7 @@ package org.apache.pinot.broker.api;
 import java.util.Set;
 import org.apache.pinot.common.request.BrokerRequest;
 import org.apache.pinot.spi.auth.AuthorizationResult;
+import org.apache.pinot.spi.auth.broker.RequesterIdentity;
 import org.testng.annotations.Test;
 
 import static org.testng.Assert.assertEquals;
