+++ pinot-core/src/main/java/org/apache/pinot/server/access/AccessControl.java
@@ -21,6 +21,7 @@ package org.apache.pinot.server.access;
 import io.netty.channel.ChannelHandlerContext;
 import org.apache.pinot.spi.annotations.InterfaceAudience;
 import org.apache.pinot.spi.annotations.InterfaceStability;
+import org.apache.pinot.spi.auth.server.RequesterIdentity;
 
 
 @InterfaceAudience.Public
