+++ pinot-core/src/main/java/org/apache/pinot/server/access/GrpcRequesterIdentity.java
@@ -21,6 +21,7 @@ package org.apache.pinot.server.access;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
 import java.util.Map;
+import org.apache.pinot.spi.auth.server.RequesterIdentity;
 
 
 /**
