+++ pinot-server/src/main/java/org/apache/pinot/server/api/resources/ServerResourceUtils.java
@@ -26,7 +26,7 @@ import org.apache.pinot.segment.local.data.manager.TableDataManager;
 import org.apache.pinot.server.access.AccessControl;
 import org.apache.pinot.server.access.AccessControlFactory;
 import org.apache.pinot.server.access.HttpRequesterIdentity;
-import org.apache.pinot.server.access.RequesterIdentity;
+import org.apache.pinot.spi.auth.server.RequesterIdentity;
 import org.apache.pinot.server.starter.ServerInstance;
